OS X XCode >= 3.1 needs a special linker flag for dynamic libraries and main programs to share common blocks. This is a required feature for the MIRIAD 'mem' routines and possibly other parts of MIRIAD as well.  We'll just blindly supply the flag if the linker accepts it.
OS X XCode = 3.0 has the default behavior of XCode 3.1, which breaks
MIRIAD, but it doesn't have the flag that allows it to be overridden.
(The change in behavior appears to have been an unintended side effect
of some other changes). MIRIAD thus appears to be un-compilable with
XCode 3.0. The culprit in particular seems to be OS X ld64 version 77.

See http://software.intel.com/en-us/forums/intel-fortran-compiler-for-linux-and-mac-os-x/topic/56995/
for related discussion.