autotools: fix wcslib build for in-place builds
I forgot to test the autotoolized wcslib build for in-place builds. It
turns out that the wcslib-provided GNUmakefile takes precedence over
automake's Makefile, leading to broken behavior by default. We work around
this in two steps: first, we set AM_MAKEFLAGS to "-f Makefile" in the
appropriate places to force the use of the correct Makefile. Second, we
write a tricksy "makedefs" so that if the user ends up attempting to
invoke the wcslib-provided GNUmakefile, it will error out.

This is definitely not ideal, but I couldn't devise a better solution
that didn't involve modifying files distributed with wcslib and tracked
in the CVS repository. Presumably vanishingly few people will need to
run a manual build in borrow/wcslib, and if so I think the error message
will be helpful enough, so I think this gets the job done satisfactorily.